+++ warden/x/intent/types/tx.pb.go
@@ -7,7 +7,7 @@ import (
 	context "context"
 	fmt "fmt"
 	_ "github.com/cosmos/cosmos-proto"
-	_ "github.com/cosmos/cosmos-sdk/codec/types"
+	types "github.com/cosmos/cosmos-sdk/codec/types"
 	_ "github.com/cosmos/cosmos-sdk/types/msgservice"
 	_ "github.com/cosmos/cosmos-sdk/types/tx/amino"
 	_ "github.com/cosmos/gogoproto/gogoproto"
@@ -125,6 +125,114 @@ func (m *MsgUpdateParamsResponse) XXX_DiscardUnknown() {
 
 var xxx_messageInfo_MsgUpdateParamsResponse proto.InternalMessageInfo
 
+type MsgNewAction struct {
+	// creator is the address of the requester of this action.
+	Creator string `protobuf:"bytes,1,opt,name=creator,proto3" json:"creator,omitempty"`
+	// message is the arbitrary message to be executed when the action is ready.
+	Message *types.Any `protobuf:"bytes,2,opt,name=message,proto3" json:"message,omitempty"`
+	// action_timeout_height is the block height up until this action can be executed.
+	ActionTimeoutHeight uint64 `protobuf:"varint,3,opt,name=action_timeout_height,json=actionTimeoutHeight,proto3" json:"action_timeout_height,omitempty"`
+}
+
+func (m *MsgNewAction) Reset()         { *m = MsgNewAction{} }
+func (m *MsgNewAction) String() string { return proto.CompactTextString(m) }
+func (*MsgNewAction) ProtoMessage()    {}
+func (*MsgNewAction) Descriptor() ([]byte, []int) {
+	return fileDescriptor_534cc1e54bd9bc1a, []int{2}
+}
+func (m *MsgNewAction) XXX_Unmarshal(b []byte) error {
+	return m.Unmarshal(b)
+}
+func (m *MsgNewAction) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
+	if deterministic {
+		return xxx_messageInfo_MsgNewAction.Marshal(b, m, deterministic)
+	} else {
+		b = b[:cap(b)]
+		n, err := m.MarshalToSizedBuffer(b)
+		if err != nil {
+			return nil, err
+		}
+		return b[:n], nil
+	}
+}
+func (m *MsgNewAction) XXX_Merge(src proto.Message) {
+	xxx_messageInfo_MsgNewAction.Merge(m, src)
+}
+func (m *MsgNewAction) XXX_Size() int {
+	return m.Size()
+}
+func (m *MsgNewAction) XXX_DiscardUnknown() {
+	xxx_messageInfo_MsgNewAction.DiscardUnknown(m)
+}
+
+var xxx_messageInfo_MsgNewAction proto.InternalMessageInfo
+
+func (m *MsgNewAction) GetCreator() string {
+	if m != nil {
+		return m.Creator
+	}
+	return ""
+}
+
+func (m *MsgNewAction) GetMessage() *types.Any {
+	if m != nil {
+		return m.Message
+	}
+	return nil
+}
+
+func (m *MsgNewAction) GetActionTimeoutHeight() uint64 {
+	if m != nil {
+		return m.ActionTimeoutHeight
+	}
+	return 0
+}
+
+type MsgNewActionResponse struct {
+	// id is the unique id of the action.
+	Id uint64 `protobuf:"varint,1,opt,name=id,proto3" json:"id,omitempty"`
+}
+
+func (m *MsgNewActionResponse) Reset()         { *m = MsgNewActionResponse{} }
+func (m *MsgNewActionResponse) String() string { return proto.CompactTextString(m) }
+func (*MsgNewActionResponse) ProtoMessage()    {}
+func (*MsgNewActionResponse) Descriptor() ([]byte, []int) {
+	return fileDescriptor_534cc1e54bd9bc1a, []int{3}
+}
+func (m *MsgNewActionResponse) XXX_Unmarshal(b []byte) error {
+	return m.Unmarshal(b)
+}
+func (m *MsgNewActionResponse) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
+	if deterministic {
+		return xxx_messageInfo_MsgNewActionResponse.Marshal(b, m, deterministic)
+	} else {
+		b = b[:cap(b)]
+		n, err := m.MarshalToSizedBuffer(b)
+		if err != nil {
+			return nil, err
+		}
+		return b[:n], nil
+	}
+}
+func (m *MsgNewActionResponse) XXX_Merge(src proto.Message) {
+	xxx_messageInfo_MsgNewActionResponse.Merge(m, src)
+}
+func (m *MsgNewActionResponse) XXX_Size() int {
+	return m.Size()
+}
+func (m *MsgNewActionResponse) XXX_DiscardUnknown() {
+	xxx_messageInfo_MsgNewActionResponse.DiscardUnknown(m)
+}
+
+var xxx_messageInfo_MsgNewActionResponse proto.InternalMessageInfo
+
+func (m *MsgNewActionResponse) GetId() uint64 {
+	if m != nil {
+		return m.Id
+	}
+	return 0
+}
+
 type MsgApproveAction struct {
 	Creator    string `protobuf:"bytes,1,opt,name=creator,proto3" json:"creator,omitempty"`
 	ActionType string `protobuf:"bytes,2,opt,name=action_type,json=actionType,proto3" json:"action_type,omitempty"`
@@ -135,7 +243,7 @@ func (m *MsgApproveAction) Reset()         { *m = MsgApproveAction{} }
 func (m *MsgApproveAction) String() string { return proto.CompactTextString(m) }
 func (*MsgApproveAction) ProtoMessage()    {}
 func (*MsgApproveAction) Descriptor() ([]byte, []int) {
-	return fileDescriptor_534cc1e54bd9bc1a, []int{2}
+	return fileDescriptor_534cc1e54bd9bc1a, []int{4}
 }
 func (m *MsgApproveAction) XXX_Unmarshal(b []byte) error {
 	return m.Unmarshal(b)
@@ -193,7 +301,7 @@ func (m *MsgApproveActionResponse) Reset()         { *m = MsgApproveActionRespon
 func (m *MsgApproveActionResponse) String() string { return proto.CompactTextString(m) }
 func (*MsgApproveActionResponse) ProtoMessage()    {}
 func (*MsgApproveActionResponse) Descriptor() ([]byte, []int) {
-	return fileDescriptor_534cc1e54bd9bc1a, []int{3}
+	return fileDescriptor_534cc1e54bd9bc1a, []int{5}
 }
 func (m *MsgApproveActionResponse) XXX_Unmarshal(b []byte) error {
 	return m.Unmarshal(b)
@@ -239,7 +347,7 @@ func (m *MsgNewIntent) Reset()         { *m = MsgNewIntent{} }
 func (m *MsgNewIntent) String() string { return proto.CompactTextString(m) }
 func (*MsgNewIntent) ProtoMessage()    {}
 func (*MsgNewIntent) Descriptor() ([]byte, []int) {
-	return fileDescriptor_534cc1e54bd9bc1a, []int{4}
+	return fileDescriptor_534cc1e54bd9bc1a, []int{6}
 }
 func (m *MsgNewIntent) XXX_Unmarshal(b []byte) error {
 	return m.Unmarshal(b)
@@ -297,7 +405,7 @@ func (m *MsgNewIntentResponse) Reset()         { *m = MsgNewIntentResponse{} }
 func (m *MsgNewIntentResponse) String() string { return proto.CompactTextString(m) }
 func (*MsgNewIntentResponse) ProtoMessage()    {}
 func (*MsgNewIntentResponse) Descriptor() ([]byte, []int) {
-	return fileDescriptor_534cc1e54bd9bc1a, []int{5}
+	return fileDescriptor_534cc1e54bd9bc1a, []int{7}
 }
 func (m *MsgNewIntentResponse) XXX_Unmarshal(b []byte) error {
 	return m.Unmarshal(b)
@@ -344,7 +452,7 @@ func (m *MsgUpdateIntent) Reset()         { *m = MsgUpdateIntent{} }
 func (m *MsgUpdateIntent) String() string { return proto.CompactTextString(m) }
 func (*MsgUpdateIntent) ProtoMessage()    {}
 func (*MsgUpdateIntent) Descriptor() ([]byte, []int) {
-	return fileDescriptor_534cc1e54bd9bc1a, []int{6}
+	return fileDescriptor_534cc1e54bd9bc1a, []int{8}
 }
 func (m *MsgUpdateIntent) XXX_Unmarshal(b []byte) error {
 	return m.Unmarshal(b)
@@ -408,7 +516,7 @@ func (m *MsgUpdateIntentResponse) Reset()         { *m = MsgUpdateIntentResponse
 func (m *MsgUpdateIntentResponse) String() string { return proto.CompactTextString(m) }
 func (*MsgUpdateIntentResponse) ProtoMessage()    {}
 func (*MsgUpdateIntentResponse) Descriptor() ([]byte, []int) {
-	return fileDescriptor_534cc1e54bd9bc1a, []int{7}
+	return fileDescriptor_534cc1e54bd9bc1a, []int{9}
 }
 func (m *MsgUpdateIntentResponse) XXX_Unmarshal(b []byte) error {
 	return m.Unmarshal(b)
@@ -447,7 +555,7 @@ func (m *MsgRevokeAction) Reset()         { *m = MsgRevokeAction{} }
 func (m *MsgRevokeAction) String() string { return proto.CompactTextString(m) }
 func (*MsgRevokeAction) ProtoMessage()    {}
 func (*MsgRevokeAction) Descriptor() ([]byte, []int) {
-	return fileDescriptor_534cc1e54bd9bc1a, []int{8}
+	return fileDescriptor_534cc1e54bd9bc1a, []int{10}
 }
 func (m *MsgRevokeAction) XXX_Unmarshal(b []byte) error {
 	return m.Unmarshal(b)
@@ -504,7 +612,7 @@ func (m *MsgRevokeActionResponse) Reset()         { *m = MsgRevokeActionResponse
 func (m *MsgRevokeActionResponse) String() string { return proto.CompactTextString(m) }
 func (*MsgRevokeActionResponse) ProtoMessage()    {}
 func (*MsgRevokeActionResponse) Descriptor() ([]byte, []int) {
-	return fileDescriptor_534cc1e54bd9bc1a, []int{9}
+	return fileDescriptor_534cc1e54bd9bc1a, []int{11}
 }
 func (m *MsgRevokeActionResponse) XXX_Unmarshal(b []byte) error {
 	return m.Unmarshal(b)
@@ -536,6 +644,8 @@ var xxx_messageInfo_MsgRevokeActionResponse proto.InternalMessageInfo
 func init() {
 	proto.RegisterType((*MsgUpdateParams)(nil), "warden.intent.MsgUpdateParams")
 	proto.RegisterType((*MsgUpdateParamsResponse)(nil), "warden.intent.MsgUpdateParamsResponse")
+	proto.RegisterType((*MsgNewAction)(nil), "warden.intent.MsgNewAction")
+	proto.RegisterType((*MsgNewActionResponse)(nil), "warden.intent.MsgNewActionResponse")
 	proto.RegisterType((*MsgApproveAction)(nil), "warden.intent.MsgApproveAction")
 	proto.RegisterType((*MsgApproveActionResponse)(nil), "warden.intent.MsgApproveActionResponse")
 	proto.RegisterType((*MsgNewIntent)(nil), "warden.intent.MsgNewIntent")
@@ -549,46 +659,51 @@ func init() {
 func init() { proto.RegisterFile("warden/intent/tx.proto", fileDescriptor_534cc1e54bd9bc1a) }
 
 var fileDescriptor_534cc1e54bd9bc1a = []byte{
-	// 621 bytes of a gzipped FileDescriptorProto
-	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0xbc, 0x94, 0xbd, 0x6f, 0xd3, 0x4e,
-	0x18, 0xc7, 0xe3, 0x24, 0xbf, 0xfe, 0xf0, 0x35, 0xe5, 0xc5, 0x0a, 0xad, 0xeb, 0x4a, 0x4e, 0x64,
-	0xa4, 0x10, 0x45, 0x6a, 0x2c, 0x82, 0x84, 0x50, 0x17, 0x94, 0x6c, 0x1d, 0x82, 0x90, 0x79, 0x91,
-	0xca, 0x52, 0x39, 0xf1, 0xd5, 0x35, 0xd4, 0x3e, 0xcb, 0x77, 0x79, 0x1b, 0x90, 0x10, 0x23, 0x13,
-	0x7f, 0x06, 0x63, 0x06, 0x24, 0x36, 0xe6, 0x8e, 0x15, 0x13, 0x13, 0x42, 0xc9, 0x90, 0x7f, 0x03,
-	0xf9, 0xee, 0xec, 0xd8, 0x6e, 0x9a, 0x6c, 0x2c, 0x89, 0x9f, 0xe7, 0xfb, 0xdc, 0xf7, 0x3e, 0xe7,
-	0xe7, 0xf1, 0x81, 0xdd, 0x91, 0x19, 0x58, 0xd0, 0xd3, 0x1d, 0x8f, 0x40, 0x8f, 0xe8, 0x64, 0xdc,
-	0xf4, 0x03, 0x44, 0x90, 0xb4, 0xc3, 0xf2, 0x4d, 0x96, 0x57, 0xee, 0x99, 0xae, 0xe3, 0x21, 0x9d,
-	0xfe, 0xb2, 0x0a, 0x65, 0xaf, 0x8f, 0xb0, 0x8b, 0xb0, 0xee, 0x62, 0x5b, 0x1f, 0x3e, 0x0a, 0xff,
-	0xb8, 0xb0, 0xcf, 0x84, 0x53, 0x1a, 0xe9, 0x2c, 0xe0, 0x52, 0xd9, 0x46, 0x36, 0x62, 0xf9, 0xf0,
-	0x29, 0x5a, 0x60, 0x23, 0x64, 0x5f, 0x40, 0x9d, 0x46, 0xbd, 0xc1, 0x99, 0x6e, 0x7a, 0x13, 0x2e,
-	0x29, 0x69, 0x3c, 0xdf, 0x0c, 0x4c, 0x97, 0x9b, 0x69, 0xdf, 0x05, 0x70, 0xa7, 0x8b, 0xed, 0xd7,
-	0xbe, 0x65, 0x12, 0xf8, 0x82, 0x2a, 0xd2, 0x13, 0x20, 0x9a, 0x03, 0x72, 0x8e, 0x02, 0x87, 0x4c,
-	0x64, 0xa1, 0x2a, 0xd4, 0xc5, 0x8e, 0xfc, 0xf3, 0xdb, 0x61, 0x99, 0x53, 0xb4, 0x2d, 0x2b, 0x80,
-	0x18, 0xbf, 0x24, 0x81, 0xe3, 0xd9, 0xc6, 0xb2, 0x54, 0x7a, 0x0a, 0xb6, 0x98, 0xb7, 0x9c, 0xaf,
-	0x0a, 0xf5, 0xed, 0xd6, 0xfd, 0x66, 0xea, 0xfc, 0x4d, 0x66, 0xdf, 0x11, 0x2f, 0x7f, 0x57, 0x72,
-	0x5f, 0x17, 0xd3, 0x86, 0x60, 0xf0, 0xfa, 0xa3, 0xd6, 0xa7, 0xc5, 0xb4, 0xb1, 0x74, 0xfa, 0xbc,
-	0x98, 0x36, 0x2a, 0x1c, 0x7a, 0x1c, 0x61, 0x67, 0x28, 0xb5, 0x7d, 0xb0, 0x97, 0x49, 0x19, 0x10,
-	0xfb, 0xc8, 0xc3, 0x50, 0x1b, 0x83, 0xbb, 0x5d, 0x6c, 0xb7, 0x7d, 0x3f, 0x40, 0x43, 0xd8, 0xee,
-	0x13, 0x07, 0x79, 0x92, 0x0c, 0xfe, 0xef, 0x07, 0xd0, 0x24, 0x28, 0x60, 0x47, 0x32, 0xa2, 0x50,
-	0xaa, 0x80, 0x6d, 0x93, 0xd6, 0x9c, 0x92, 0x89, 0x0f, 0x29, 0xbb, 0x68, 0x00, 0x96, 0x7a, 0x35,
-	0xf1, 0xa1, 0x74, 0x00, 0x44, 0x5e, 0xe0, 0x58, 0x72, 0xa1, 0x2a, 0xd4, 0x8b, 0xc6, 0x2d, 0x96,
-	0x38, 0xb6, 0x8e, 0x4a, 0x21, 0x7a, 0xe4, 0xa5, 0xb5, 0x80, 0x9c, 0xdd, 0x39, 0xa2, 0x92, 0x76,
-	0xc1, 0x16, 0x26, 0x26, 0x19, 0x60, 0x0e, 0xc0, 0x23, 0xed, 0x1d, 0x28, 0x75, 0xb1, 0xfd, 0x1c,
-	0x8e, 0x8e, 0xe9, 0x41, 0xd7, 0x90, 0x4a, 0xa0, 0xe8, 0x99, 0x6e, 0x84, 0x48, 0x9f, 0x25, 0x15,
-	0x00, 0x0b, 0x9e, 0x39, 0x9e, 0x13, 0xee, 0x45, 0xe9, 0x44, 0x23, 0x91, 0xc9, 0xf0, 0xd5, 0x40,
-	0x39, 0xb9, 0x57, 0xcc, 0x76, 0x1b, 0xe4, 0x1d, 0x8b, 0x6e, 0x57, 0x34, 0xf2, 0x8e, 0xa5, 0x7d,
-	0x48, 0x4c, 0xc5, 0x46, 0x2c, 0xb6, 0x38, 0x1f, 0x2d, 0x8e, 0x31, 0x0b, 0x37, 0x62, 0x16, 0x37,
-	0x60, 0x26, 0x7b, 0x9b, 0x26, 0xd5, 0x46, 0x94, 0xcc, 0x80, 0x43, 0xf4, 0xfe, 0xdf, 0xb6, 0x96,
-	0x31, 0x25, 0x37, 0x8e, 0x98, 0x5a, 0x3f, 0x0a, 0xa0, 0xd0, 0xc5, 0xb6, 0xf4, 0x06, 0x94, 0x52,
-	0x1f, 0x92, 0x9a, 0xf9, 0x00, 0x32, 0xf3, 0xaa, 0xd4, 0xd6, 0xeb, 0x71, 0x77, 0x4e, 0xc0, 0x4e,
-	0x7a, 0x98, 0x2b, 0xd7, 0x17, 0xa6, 0x0a, 0x94, 0x87, 0x1b, 0x0a, 0x62, 0xeb, 0x2e, 0x10, 0x97,
-	0x93, 0x77, 0x70, 0x7d, 0x55, 0x2c, 0x2a, 0x0f, 0xd6, 0x88, 0xb1, 0x5d, 0xfc, 0x06, 0xb8, 0xe3,
-	0x8d, 0x6f, 0x80, 0x9b, 0xd6, 0xd6, 0xeb, 0x49, 0xdf, 0x54, 0xcb, 0x57, 0xf8, 0x26, 0xf5, 0x55,
-	0xbe, 0xab, 0x3a, 0xa7, 0xfc, 0xf7, 0x31, 0xbc, 0x87, 0x3a, 0x27, 0x97, 0x33, 0x55, 0xb8, 0x9a,
-	0xa9, 0xc2, 0x9f, 0x99, 0x2a, 0x7c, 0x99, 0xab, 0xb9, 0xab, 0xb9, 0x9a, 0xfb, 0x35, 0x57, 0x73,
-	0x6f, 0x9f, 0xd9, 0x0e, 0x39, 0x1f, 0xf4, 0x9a, 0x7d, 0xe4, 0xea, 0xcc, 0xf2, 0x90, 0x5e, 0x9c,
-	0x7d, 0x74, 0xc1, 0xe3, 0x4c, 0xb8, 0xbc, 0xb0, 0xc2, 0x91, 0xc3, 0xbd, 0x2d, 0xaa, 0x3f, 0xfe,
-	0x1b, 0x00, 0x00, 0xff, 0xff, 0x86, 0x5e, 0x0d, 0xf5, 0x24, 0x06, 0x00, 0x00,
+	// 692 bytes of a gzipped FileDescriptorProto
+	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0xbc, 0x94, 0x4d, 0x4f, 0xd4, 0x40,
+	0x18, 0xc7, 0xb7, 0xcb, 0x0a, 0x76, 0x00, 0x5f, 0xea, 0x02, 0x4b, 0x49, 0xba, 0xa4, 0x26, 0x48,
+	0x48, 0x68, 0xe3, 0x9a, 0x18, 0xc3, 0xc5, 0x2c, 0x27, 0x39, 0xac, 0x31, 0xf5, 0x25, 0xc1, 0x0b,
+	0x29, 0xdb, 0x61, 0x76, 0x94, 0x76, 0x9a, 0xce, 0x2c, 0xb0, 0x07, 0x13, 0xe3, 0xd1, 0xc4, 0xc4,
+	0xb3, 0x9f, 0xc0, 0x23, 0x07, 0x13, 0xbf, 0x02, 0x47, 0xe2, 0xc9, 0x93, 0x31, 0x70, 0xe0, 0x6b,
+	0x98, 0x9d, 0x97, 0x6e, 0x5b, 0xf6, 0xe5, 0xe6, 0x65, 0xb7, 0xcf, 0xfc, 0x9f, 0xf9, 0xcf, 0x6f,
+	0x9e, 0x79, 0x66, 0xc0, 0xe2, 0xb1, 0x9f, 0x04, 0x30, 0x72, 0x71, 0xc4, 0x60, 0xc4, 0x5c, 0x76,
+	0xe2, 0xc4, 0x09, 0x61, 0xc4, 0x98, 0x17, 0xe3, 0x8e, 0x18, 0x37, 0xef, 0xfa, 0x21, 0x8e, 0x88,
+	0xcb, 0x7f, 0x45, 0x86, 0xb9, 0xd4, 0x26, 0x34, 0x24, 0xd4, 0x0d, 0x29, 0x72, 0x8f, 0x1e, 0xf6,
+	0xff, 0xa4, 0xb0, 0x2c, 0x84, 0x3d, 0x1e, 0xb9, 0x22, 0x90, 0x52, 0x15, 0x11, 0x44, 0xc4, 0x78,
+	0xff, 0x4b, 0x4d, 0x40, 0x84, 0xa0, 0x43, 0xe8, 0xf2, 0x68, 0xbf, 0x7b, 0xe0, 0xfa, 0x51, 0x4f,
+	0x4a, 0x66, 0x1e, 0x2f, 0xf6, 0x13, 0x3f, 0x94, 0x66, 0xf6, 0x4f, 0x0d, 0xdc, 0x6e, 0x51, 0xf4,
+	0x3a, 0x0e, 0x7c, 0x06, 0x5f, 0x70, 0xc5, 0x78, 0x0c, 0x74, 0xbf, 0xcb, 0x3a, 0x24, 0xc1, 0xac,
+	0x57, 0xd3, 0x56, 0xb5, 0x75, 0x7d, 0xbb, 0xf6, 0xeb, 0xc7, 0x66, 0x55, 0x52, 0x34, 0x83, 0x20,
+	0x81, 0x94, 0xbe, 0x64, 0x09, 0x8e, 0x90, 0x37, 0x48, 0x35, 0x9e, 0x80, 0x69, 0xe1, 0x5d, 0x2b,
+	0xaf, 0x6a, 0xeb, 0xb3, 0x8d, 0x05, 0x27, 0xb7, 0x7f, 0x47, 0xd8, 0x6f, 0xeb, 0x67, 0x7f, 0xea,
+	0xa5, 0xef, 0x57, 0xa7, 0x1b, 0x9a, 0x27, 0xf3, 0xb7, 0x1a, 0x9f, 0xae, 0x4e, 0x37, 0x06, 0x4e,
+	0x9f, 0xaf, 0x4e, 0x37, 0xea, 0x12, 0xfa, 0x44, 0x61, 0x17, 0x28, 0xed, 0x65, 0xb0, 0x54, 0x18,
+	0xf2, 0x20, 0x8d, 0x49, 0x44, 0xa1, 0xfd, 0x4d, 0x03, 0x73, 0x2d, 0x8a, 0x9e, 0xc3, 0xe3, 0x66,
+	0x9b, 0x61, 0x12, 0x19, 0x35, 0x30, 0xd3, 0x4e, 0xa0, 0xcf, 0x48, 0x22, 0xf6, 0xe3, 0xa9, 0xd0,
+	0x70, 0xc0, 0x4c, 0x08, 0x29, 0xf5, 0x11, 0x94, 0xd0, 0x55, 0x47, 0x14, 0xd2, 0x51, 0x85, 0x74,
+	0x9a, 0x51, 0xcf, 0x53, 0x49, 0x46, 0x03, 0x2c, 0xf8, 0xdc, 0x73, 0x8f, 0xe1, 0x10, 0x92, 0x2e,
+	0xdb, 0xeb, 0x40, 0x8c, 0x3a, 0xac, 0x36, 0xb5, 0xaa, 0xad, 0x57, 0xbc, 0x7b, 0x42, 0x7c, 0x25,
+	0xb4, 0x67, 0x5c, 0xda, 0x9a, 0xeb, 0xef, 0x4e, 0xad, 0x68, 0xaf, 0x81, 0x6a, 0x96, 0x4d, 0x41,
+	0x1b, 0xb7, 0x40, 0x19, 0x07, 0x1c, 0xaf, 0xe2, 0x95, 0x71, 0x60, 0x9f, 0x80, 0x3b, 0x2d, 0x8a,
+	0x9a, 0x71, 0x9c, 0x90, 0x23, 0x38, 0x71, 0x1f, 0x75, 0x30, 0xab, 0xb8, 0x7a, 0xb1, 0xd8, 0x8b,
+	0xee, 0x01, 0x49, 0xd3, 0x8b, 0xa1, 0xb1, 0x02, 0x74, 0x99, 0x80, 0x03, 0x09, 0x7b, 0x53, 0x0c,
+	0xec, 0x04, 0x05, 0xc2, 0x06, 0xa8, 0x15, 0x57, 0x4e, 0x29, 0x17, 0xc1, 0x34, 0x65, 0x3e, 0xeb,
+	0x52, 0x09, 0x20, 0x23, 0xfb, 0x9d, 0xaa, 0xf8, 0x0e, 0x3f, 0xad, 0x31, 0xa4, 0x06, 0xa8, 0x44,
+	0x7e, 0xa8, 0x10, 0xf9, 0xb7, 0x61, 0x01, 0x10, 0xc0, 0x03, 0x1c, 0xe1, 0xfe, 0x5a, 0x9c, 0x4e,
+	0xf7, 0x32, 0x23, 0xa3, 0x2a, 0x28, 0xd6, 0x1a, 0x59, 0xc1, 0x0f, 0x99, 0xd6, 0x9e, 0x88, 0x25,
+	0x26, 0x97, 0xd5, 0xe4, 0x14, 0x73, 0x6a, 0x24, 0x66, 0x65, 0x02, 0x66, 0xb6, 0x41, 0xf3, 0xa4,
+	0xf6, 0x31, 0x27, 0xf3, 0xe0, 0x11, 0x79, 0xff, 0x7f, 0x8f, 0x56, 0x30, 0x65, 0x17, 0x56, 0x4c,
+	0x8d, 0x2f, 0x15, 0x30, 0xd5, 0xa2, 0xc8, 0x78, 0x03, 0xe6, 0x72, 0xaf, 0x81, 0x55, 0xb8, 0xc5,
+	0x85, 0x4b, 0x67, 0xae, 0x8d, 0xd7, 0xd3, 0xd3, 0x69, 0x01, 0x7d, 0x70, 0x21, 0x57, 0xae, 0x4f,
+	0x4a, 0x45, 0xf3, 0xfe, 0x18, 0x31, 0xb5, 0xdb, 0x05, 0xf3, 0xf9, 0xbb, 0x51, 0xbf, 0x3e, 0x2b,
+	0x97, 0x60, 0x3e, 0x98, 0x90, 0x50, 0x20, 0x95, 0x1d, 0x33, 0x9c, 0x54, 0x88, 0x23, 0x48, 0x0b,
+	0x6d, 0x99, 0x16, 0x54, 0x3a, 0x8e, 0x2c, 0xa8, 0x34, 0x5d, 0x1b, 0xaf, 0x67, 0x7d, 0x73, 0x1d,
+	0x34, 0xc4, 0x37, 0xab, 0x0f, 0xf3, 0x1d, 0xd6, 0x08, 0xe6, 0x8d, 0x8f, 0xfd, 0xb7, 0x79, 0x7b,
+	0xf7, 0xec, 0xc2, 0xd2, 0xce, 0x2f, 0x2c, 0xed, 0xef, 0x85, 0xa5, 0x7d, 0xbd, 0xb4, 0x4a, 0xe7,
+	0x97, 0x56, 0xe9, 0xf7, 0xa5, 0x55, 0x7a, 0xfb, 0x14, 0x61, 0xd6, 0xe9, 0xee, 0x3b, 0x6d, 0x12,
+	0xba, 0xc2, 0x72, 0x93, 0x3f, 0x96, 0x6d, 0x72, 0x28, 0xe3, 0x42, 0x38, 0x78, 0xc4, 0xfb, 0x1d,
+	0x4c, 0xf7, 0xa7, 0xb9, 0xfe, 0xe8, 0x5f, 0x00, 0x00, 0x00, 0xff, 0xff, 0x64, 0x20, 0x93, 0x95,
+	0x38, 0x07, 0x00, 0x00,
 }
 
 // Reference imports to suppress errors if they are not otherwise used.
@@ -606,6 +721,8 @@ type MsgClient interface {
 	// UpdateParams defines a (governance) operation for updating the module
 	// parameters. The authority defaults to the x/gov module account.
 	UpdateParams(ctx context.Context, in *MsgUpdateParams, opts ...grpc.CallOption) (*MsgUpdateParamsResponse, error)
+	// NewAction creates a new Action.
+	NewAction(ctx context.Context, in *MsgNewAction, opts ...grpc.CallOption) (*MsgNewActionResponse, error)
 	// Add an approval to an existing Action.
 	ApproveAction(ctx context.Context, in *MsgApproveAction, opts ...grpc.CallOption) (*MsgApproveActionResponse, error)
 	// Create a new intent.
@@ -627,6 +744,15 @@ func NewMsgClient(cc grpc1.ClientConn) MsgClient {
 func (c *msgClient) UpdateParams(ctx context.Context, in *MsgUpdateParams, opts ...grpc.CallOption) (*MsgUpdateParamsResponse, error) {
 	out := new(MsgUpdateParamsResponse)
 	err := c.cc.Invoke(ctx, "/warden.intent.Msg/UpdateParams", in, out, opts...)
+	if err != nil {
+		return nil, err
+	}
+	return out, nil
+}
+
+func (c *msgClient) NewAction(ctx context.Context, in *MsgNewAction, opts ...grpc.CallOption) (*MsgNewActionResponse, error) {
+	out := new(MsgNewActionResponse)
+	err := c.cc.Invoke(ctx, "/warden.intent.Msg/NewAction", in, out, opts...)
 	if err != nil {
 		return nil, err
 	}
@@ -674,6 +800,8 @@ type MsgServer interface {
 	// UpdateParams defines a (governance) operation for updating the module
 	// parameters. The authority defaults to the x/gov module account.
 	UpdateParams(context.Context, *MsgUpdateParams) (*MsgUpdateParamsResponse, error)
+	// NewAction creates a new Action.
+	NewAction(context.Context, *MsgNewAction) (*MsgNewActionResponse, error)
 	// Add an approval to an existing Action.
 	ApproveAction(context.Context, *MsgApproveAction) (*MsgApproveActionResponse, error)
 	// Create a new intent.
@@ -690,6 +818,9 @@ type UnimplementedMsgServer struct {
 
 func (*UnimplementedMsgServer) UpdateParams(ctx context.Context, req *MsgUpdateParams) (*MsgUpdateParamsResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method UpdateParams not implemented")
+}
+func (*UnimplementedMsgServer) NewAction(ctx context.Context, req *MsgNewAction) (*MsgNewActionResponse, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method NewAction not implemented")
 }
 func (*UnimplementedMsgServer) ApproveAction(ctx context.Context, req *MsgApproveAction) (*MsgApproveActionResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method ApproveAction not implemented")
@@ -722,6 +853,24 @@ func _Msg_UpdateParams_Handler(srv interface{}, ctx context.Context, dec func(in
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(MsgServer).UpdateParams(ctx, req.(*MsgUpdateParams))
+	}
+	return interceptor(ctx, in, info, handler)
+}
+
+func _Msg_NewAction_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(MsgNewAction)
+	if err := dec(in); err != nil {
+		return nil, err
+	}
+	if interceptor == nil {
+		return srv.(MsgServer).NewAction(ctx, in)
+	}
+	info := &grpc.UnaryServerInfo{
+		Server:     srv,
+		FullMethod: "/warden.intent.Msg/NewAction",
+	}
+	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
+		return srv.(MsgServer).NewAction(ctx, req.(*MsgNewAction))
 	}
 	return interceptor(ctx, in, info, handler)
 }
@@ -807,6 +956,10 @@ var _Msg_serviceDesc = grpc.ServiceDesc{
 			Handler:    _Msg_UpdateParams_Handler,
 		},
 		{
+			MethodName: "NewAction",
+			Handler:    _Msg_NewAction_Handler,
+		},
+		{
 			MethodName: "ApproveAction",
 			Handler:    _Msg_ApproveAction_Handler,
 		},
@@ -887,6 +1040,81 @@ func (m *MsgUpdateParamsResponse) MarshalToSizedBuffer(dAtA []byte) (int, error)
 	_ = i
 	var l int
 	_ = l
+	return len(dAtA) - i, nil
+}
+
+func (m *MsgNewAction) Marshal() (dAtA []byte, err error) {
+	size := m.Size()
+	dAtA = make([]byte, size)
+	n, err := m.MarshalToSizedBuffer(dAtA[:size])
+	if err != nil {
+		return nil, err
+	}
+	return dAtA[:n], nil
+}
+
+func (m *MsgNewAction) MarshalTo(dAtA []byte) (int, error) {
+	size := m.Size()
+	return m.MarshalToSizedBuffer(dAtA[:size])
+}
+
+func (m *MsgNewAction) MarshalToSizedBuffer(dAtA []byte) (int, error) {
+	i := len(dAtA)
+	_ = i
+	var l int
+	_ = l
+	if m.ActionTimeoutHeight != 0 {
+		i = encodeVarintTx(dAtA, i, uint64(m.ActionTimeoutHeight))
+		i--
+		dAtA[i] = 0x18
+	}
+	if m.Message != nil {
+		{
+			size, err := m.Message.MarshalToSizedBuffer(dAtA[:i])
+			if err != nil {
+				return 0, err
+			}
+			i -= size
+			i = encodeVarintTx(dAtA, i, uint64(size))
+		}
+		i--
+		dAtA[i] = 0x12
+	}
+	if len(m.Creator) > 0 {
+		i -= len(m.Creator)
+		copy(dAtA[i:], m.Creator)
+		i = encodeVarintTx(dAtA, i, uint64(len(m.Creator)))
+		i--
+		dAtA[i] = 0xa
+	}
+	return len(dAtA) - i, nil
+}
+
+func (m *MsgNewActionResponse) Marshal() (dAtA []byte, err error) {
+	size := m.Size()
+	dAtA = make([]byte, size)
+	n, err := m.MarshalToSizedBuffer(dAtA[:size])
+	if err != nil {
+		return nil, err
+	}
+	return dAtA[:n], nil
+}
+
+func (m *MsgNewActionResponse) MarshalTo(dAtA []byte) (int, error) {
+	size := m.Size()
+	return m.MarshalToSizedBuffer(dAtA[:size])
+}
+
+func (m *MsgNewActionResponse) MarshalToSizedBuffer(dAtA []byte) (int, error) {
+	i := len(dAtA)
+	_ = i
+	var l int
+	_ = l
+	if m.Id != 0 {
+		i = encodeVarintTx(dAtA, i, uint64(m.Id))
+		i--
+		dAtA[i] = 0x8
+	}
 	return len(dAtA) - i, nil
 }
 
@@ -1206,6 +1434,38 @@ func (m *MsgUpdateParamsResponse) Size() (n int) {
 	return n
 }
 
+func (m *MsgNewAction) Size() (n int) {
+	if m == nil {
+		return 0
+	}
+	var l int
+	_ = l
+	l = len(m.Creator)
+	if l > 0 {
+		n += 1 + l + sovTx(uint64(l))
+	}
+	if m.Message != nil {
+		l = m.Message.Size()
+		n += 1 + l + sovTx(uint64(l))
+	}
+	if m.ActionTimeoutHeight != 0 {
+		n += 1 + sovTx(uint64(m.ActionTimeoutHeight))
+	}
+	return n
+}
+
+func (m *MsgNewActionResponse) Size() (n int) {
+	if m == nil {
+		return 0
+	}
+	var l int
+	_ = l
+	if m.Id != 0 {
+		n += 1 + sovTx(uint64(m.Id))
+	}
+	return n
+}
+
 func (m *MsgApproveAction) Size() (n int) {
 	if m == nil {
 		return 0
@@ -1484,6 +1744,212 @@ func (m *MsgUpdateParamsResponse) Unmarshal(dAtA []byte) error {
 			return fmt.Errorf("proto: MsgUpdateParamsResponse: illegal tag %d (wire type %d)", fieldNum, wire)
 		}
 		switch fieldNum {
+		default:
+			iNdEx = preIndex
+			skippy, err := skipTx(dAtA[iNdEx:])
+			if err != nil {
+				return err
+			}
+			if (skippy < 0) || (iNdEx+skippy) < 0 {
+				return ErrInvalidLengthTx
+			}
+			if (iNdEx + skippy) > l {
+				return io.ErrUnexpectedEOF
+			}
+			iNdEx += skippy
+		}
+	}
+
+	if iNdEx > l {
+		return io.ErrUnexpectedEOF
+	}
+	return nil
+}
+func (m *MsgNewAction) Unmarshal(dAtA []byte) error {
+	l := len(dAtA)
+	iNdEx := 0
+	for iNdEx < l {
+		preIndex := iNdEx
+		var wire uint64
+		for shift := uint(0); ; shift += 7 {
+			if shift >= 64 {
+				return ErrIntOverflowTx
+			}
+			if iNdEx >= l {
+				return io.ErrUnexpectedEOF
+			}
+			b := dAtA[iNdEx]
+			iNdEx++
+			wire |= uint64(b&0x7F) << shift
+			if b < 0x80 {
+				break
+			}
+		}
+		fieldNum := int32(wire >> 3)
+		wireType := int(wire & 0x7)
+		if wireType == 4 {
+			return fmt.Errorf("proto: MsgNewAction: wiretype end group for non-group")
+		}
+		if fieldNum <= 0 {
+			return fmt.Errorf("proto: MsgNewAction: illegal tag %d (wire type %d)", fieldNum, wire)
+		}
+		switch fieldNum {
+		case 1:
+			if wireType != 2 {
+				return fmt.Errorf("proto: wrong wireType = %d for field Creator", wireType)
+			}
+			var stringLen uint64
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowTx
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				stringLen |= uint64(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+			intStringLen := int(stringLen)
+			if intStringLen < 0 {
+				return ErrInvalidLengthTx
+			}
+			postIndex := iNdEx + intStringLen
+			if postIndex < 0 {
+				return ErrInvalidLengthTx
+			}
+			if postIndex > l {
+				return io.ErrUnexpectedEOF
+			}
+			m.Creator = string(dAtA[iNdEx:postIndex])
+			iNdEx = postIndex
+		case 2:
+			if wireType != 2 {
+				return fmt.Errorf("proto: wrong wireType = %d for field Message", wireType)
+			}
+			var msglen int
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowTx
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				msglen |= int(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+			if msglen < 0 {
+				return ErrInvalidLengthTx
+			}
+			postIndex := iNdEx + msglen
+			if postIndex < 0 {
+				return ErrInvalidLengthTx
+			}
+			if postIndex > l {
+				return io.ErrUnexpectedEOF
+			}
+			if m.Message == nil {
+				m.Message = &types.Any{}
+			}
+			if err := m.Message.Unmarshal(dAtA[iNdEx:postIndex]); err != nil {
+				return err
+			}
+			iNdEx = postIndex
+		case 3:
+			if wireType != 0 {
+				return fmt.Errorf("proto: wrong wireType = %d for field ActionTimeoutHeight", wireType)
+			}
+			m.ActionTimeoutHeight = 0
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowTx
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				m.ActionTimeoutHeight |= uint64(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+		default:
+			iNdEx = preIndex
+			skippy, err := skipTx(dAtA[iNdEx:])
+			if err != nil {
+				return err
+			}
+			if (skippy < 0) || (iNdEx+skippy) < 0 {
+				return ErrInvalidLengthTx
+			}
+			if (iNdEx + skippy) > l {
+				return io.ErrUnexpectedEOF
+			}
+			iNdEx += skippy
+		}
+	}
+
+	if iNdEx > l {
+		return io.ErrUnexpectedEOF
+	}
+	return nil
+}
+func (m *MsgNewActionResponse) Unmarshal(dAtA []byte) error {
+	l := len(dAtA)
+	iNdEx := 0
+	for iNdEx < l {
+		preIndex := iNdEx
+		var wire uint64
+		for shift := uint(0); ; shift += 7 {
+			if shift >= 64 {
+				return ErrIntOverflowTx
+			}
+			if iNdEx >= l {
+				return io.ErrUnexpectedEOF
+			}
+			b := dAtA[iNdEx]
+			iNdEx++
+			wire |= uint64(b&0x7F) << shift
+			if b < 0x80 {
+				break
+			}
+		}
+		fieldNum := int32(wire >> 3)
+		wireType := int(wire & 0x7)
+		if wireType == 4 {
+			return fmt.Errorf("proto: MsgNewActionResponse: wiretype end group for non-group")
+		}
+		if fieldNum <= 0 {
+			return fmt.Errorf("proto: MsgNewActionResponse: illegal tag %d (wire type %d)", fieldNum, wire)
+		}
+		switch fieldNum {
+		case 1:
+			if wireType != 0 {
+				return fmt.Errorf("proto: wrong wireType = %d for field Id", wireType)
+			}
+			m.Id = 0
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowTx
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				m.Id |= uint64(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
 		default:
 			iNdEx = preIndex
 			skippy, err := skipTx(dAtA[iNdEx:])